Bug 608379 - Re-Fix mail fetching with exchange 2010
When testing against exchange 2010, evolution-mapi was displaying the first
mail in my inbox but was failing to download any further mails.  The root cause
of the bug can be illustrated by the following pseudo-code:

	rowset = QueryRows;
	struct SPropTagArray tags = set_SPropTagArray(...);
	for (i = 0 to rowset.numRows) {
		msg = OpenMessage(rowset[i]);
		GetProps (msg, tags,..)
	}

so we re-use the 'tags' variable between calls to GetProps().  But the problem
is that it seems that GetProps() modifies 'tags', so subsequent calls will fail.
Creating a copy of 'tags' for each loop iteration fixes the problem and allows
me to see all messages in my inbox.