Properly fix fetching with exchange 2010
After discussing this issue with jkerihuel on IRC, we determined that my
previous patch was really just a workaround and the real problem was improper
allocation of the GetPropsTagArray variable.  We should hardly ever need to use
talloc directly and here we were using it incorrectly (which, in addition to
improper operation, also resulted in a memory leak).  This patch essentially
reverts my previous 'fix' and implements things properly.  With this patch,
things still work with exchange 2010, but the patch is much cleaner and more
correct.

Excerpts from the IRC conversation:

<kerihuel> using talloc incorrectly can sometimes lead to unpredictable behavior
   which sounds like what you've experienced
<kerihuel> it only got fixed when you've started to overwrite the existing array
...
<kerihuel> jonner: line 1170, then line 1245. You allocate the SPropTagArray
   with mem_ctx, then the sub elements with the same parent context
<kerihuel> which means that when you call MAPIFreeBuffer over GetPropsArray,
   none of the ulPropTag you allocated memory for will be released

https://bugzilla.gnome.org/show_bug.cgi?id=608379