Fix for RRULE bin parsing for recurrences with modifications
This is an additional fix for Bug #604029

This patch fixes some incorrect parsing that would cause
exchange_mapi_cal_util_bin_to_rrule() to incorrectly discard RRULE
information when processing a recurrence containing modified
instances.

Note that this does *not* include a fix to actually parse the modified
instances, which after some analysis will be fairly complicated to do and
thus saved for a later point in time.

The resulting behavior from this patch is that the recurrence will show
in the calendar, with exception dates for the modified instances, but no
corresponding events (i.e. the modified instances will be missing).