Bug #629057 - Some events created in evolution are missing in OWA
For recurring instances, an extra MAPI named property PidLidTimeZone
is required for the events to appear in OWA.  The value of this property
is an index into a rather arbitrarily defined table of timezone DST/SDT
transition rules.

Since we can not guarantee that the table will match 100% to the current
or future timezone transition rules, use a best effort heuristic to find
the closest matching rule;  for all table entries with the same UTC offset
as the current timezone, test dates before and after the transition
rules and take the entry which matches closest to what is expected.