Store last refresh information rather in folder's CamelStoreInfo
CamelFolderSummary is freed as soon as the CamelFolder, which is freed
when user moves to a different folder in the folder tree, thus storing
any information about last update during current session is useless
there. The CamelStoreInfo is persistent during the session, thus use it
to know about last folder checking. These values are not saved, they
are meant to be only for the current session.