Unref connection in a dedicated thread on EBackend::online change
When network connection changes, the backend can be moved to offline state,
but the EMapiConnection (and libmapi below it) still tries to disconnect from
the server gracefully, which can mean to wait for a connection timeout.
Rather than block whole backend, and the main thread of the factory, in this
waiting, do the unref in a dedicated thread, where it can wait as long as necessary.