Update message info's attachments flag after the message is downloaded
The server can claim in MSGFLAGS that the message has an attachment
even when it doesn't have it. This change corrects that flag after
the message is loaded from the server, in which case libcamel knows
the message structure and can determine on its own.

Related to https://gitlab.gnome.org/GNOME/evolution-ews/issues/81